Use dimensions and colour restrictions for Nokia 3310
These are based on the game jam rules:
- Screen resolution is 84x48
- Screen is scalable, by default we do x10 to make it easy to see
- Only 2 colours are allowed: Dark green & Light green

The colours are specified as a palette that can be referenced by name or
if you use it to as a real palette e.g. with an array of 0 and 1 or
similar converted from a PNG the OFF/0 pixel will be dark green and the
ON/1 pixel will be light green.