Move Nokia 3310 specific code to its own file
This way it's easier to separate your code from the code required for
the jam-specific restrictions (screen size, colours etc.).  This code
wasn't so complicated this far but adding more palettes and maybe
support for some other restrictions really warrants it being in its own
place so that whoever uses this template doesn't need to move it there
later.